Remove ResourceDataBase
Move SerializedType SerializedObject to Serialization namespace
