+++ Editor/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb6eab3cd559c9941a7d37f60889cf1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
