+++ Editor/Serialization/DynamicTypeBuilder.cs
@@ -2,7 +2,7 @@ using System;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-namespace Kurisu.Framework.Editor
+namespace Kurisu.Framework.Serialization.Editor
 {
     public static class DynamicTypeBuilder
     {
