+++ Editor/Serialization/SerializedObjectDrawer.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
 using UnityEngine;
-namespace Kurisu.Framework.Editor
+namespace Kurisu.Framework.Serialization.Editor
 {
     [CustomPropertyDrawer(typeof(SerializedObject<>))]
     public class SerializedObjectDrawer : PropertyDrawer
