+++ Runtime/Core/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a693326d3b9762443967be949a2192bf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
