+++ Runtime/Core/Serialization/SerializedObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8aa5053cfc0de134e81e2d9688375598
+guid: e317e959e39d1424b8688b32ef342e93
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
